Avanzamos

+++ server/libs/clockgame.c
@@ -5,11 +5,11 @@ void clockgame_init(clockgame_t *clockgame){
 }
 
 /* retorna el tiempo actual */
-uint64_t clockgame_time(clockgame_t *clockgame){
+uint16_t clockgame_time(clockgame_t *clockgame){
 	return *clockgame;
 }
 
 /* Incrementa el reloj en los segundos indicados */
-void clockgame_add(clockgame_t *clockgame, uint64_t seconds){
+void clockgame_add(clockgame_t *clockgame, uint16_t seconds){
 	*clockgame += seconds;
 }
