+++ server/libs/clockgame.h
@@ -7,13 +7,13 @@
 
 #include <stdint.h>
 
-typedef uint64_t clockgame_t;
+typedef uint16_t clockgame_t;
 
 /* Inicializa el reloj */
 void clockgame_init(clockgame_t *clockgame);
 /* retorna el tiempo actual */
-uint64_t clockgame_time(clockgame_t *clockgame);
+uint16_t clockgame_time(clockgame_t *clockgame);
 /* Incrementa el reloj en los segundos indicados */
-void clockgame_add(clockgame_t *clockgame, uint64_t seconds);
+void clockgame_add(clockgame_t *clockgame, uint16_t seconds);
 
 #endif
